XSL - hardenning utility-tpl.xsl
The provided PIGMA dataset sample MD will make the transformations
involved in the editor fail, because of unexpected multivalued elements.
This fixes the issue by selecting the first one.

TODO: it might be unecessary and undesirable to commit it as is, it was
just meant to be a quick workaround related to the provided sample.